Fix unit tests after rebase
