+++ extensions/route_map/test/macros.rs
@@ -3,7 +3,7 @@ macro_rules! get_routes {
     ($py:ident, $import_from:expr) => {{
         let locals = pyo3::types::PyDict::new($py);
         $py.run(include_str!($import_from), None, Some(locals))?;
-        let route = locals.get_item("route").unwrap().to_object($py);
+        let route = locals.get_item("route").unwrap();
         vec![route]
     }};
 }
